requirements-docker.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
apispec==6.4.0
apispec-webframeworks==1.0.0
arrow==1.3.0
async-timeout==4.0.3
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.34
botocore==1.34.128
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.1
cookiecutter==2.5.0
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
deepdiff==6.7.1
dnspython==2.6.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.15.1
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gevent==23.9.1
gevent-websocket==0.10.1
gitignore_parser==0.1.11
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
kthread==0.2.3
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.8
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.79
layoutparser==0.3.4
lxml==5.2.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
onnx==1.16.1
onnxruntime==1.18.0
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.5
packaging==24.1
pandas==2.2.0
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pillow==10.3.0
pillow_heif==0.16.0
portalocker==2.8.2
protobuf==5.27.1
pyarrow==15.0.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.6.1
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.2
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.9.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
s3transfer==0.10.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.25
sympy==1.12.1
tabulate==0.9.0
taipy==3.1.1
taipy-config==3.1.1
taipy-core==3.1.1
taipy-gui==3.1.2
taipy-rest==3.1.1
taipy-templates==3.1.1
tenacity==8.4.1
text-unidecode==1.3
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
Twisted==23.10.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
unstructured==0.14.6
unstructured-client==0.23.7
unstructured-inference==0.7.35
unstructured.pytesseract==0.3.12
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
